p2p networking working
